cache road path on blueprint
